Update `ModuleConflictResolver` to use the details pattern
Instead of having a single `select` method which returns the selected candidate, the method is now void, but uses
a `ConflictResolverDetails` parameter, that a conflict resolver can use to select a candidate. This enables cleaner
separation from error cases (in which case we need to call the `fail` method), from actual selection (the `select`
method). It also makes it possible to clearly separate cases where a resolver had _no opinion_ from the case where
it doesn't actually perform a selection.

This work is preliminary work to enabling resolvers that will not _choose_ a version, but instead restart part of
the resolution process based on new information.