Rework intersecting ranges handling
This commit reworks the intersection range discovery handling using the
infrastructure put in place in the previous commit: instead of having
special metadata which has to be wrapped/unwrapped using a special case,
there is now a conflict resolution rule aimed at restarting selection for
a node when we see that 2 conflicting ranges are intersecting. This can
happen when a new node is added in the graph with a different range.